+++ module/data/effectiveness-data.js
@@ -17,7 +17,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Fighting: {
             Normal: 1,
@@ -37,7 +38,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 0.5,
-            Fairy: 2
+            Fairy: 2,
+            Nuclear: 2
         },
         Flying: {
             Normal: 1,
@@ -57,7 +59,8 @@ export const TypeEffectiveness = {
             Ice: 2,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Poison: {
             Normal: 1,
@@ -77,7 +80,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 0.5
+            Fairy: 0.5,
+            Nuclear: 2
         },
         Ground: {
             Normal: 1,
@@ -97,7 +101,8 @@ export const TypeEffectiveness = {
             Ice: 2,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Rock: {
             Normal: 0.5,
@@ -117,7 +122,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Bug: {
             Normal: 1,
@@ -137,7 +143,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Ghost: {
             Normal: 0,
@@ -157,7 +164,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 2,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Steel: {
             Normal: 0.5,
@@ -177,7 +185,8 @@ export const TypeEffectiveness = {
             Ice: 0.5,
             Dragon: 0.5,
             Dark: 1,
-            Fairy: 0.5
+            Fairy: 0.5,
+            Nuclear: 0.5
         },
         Fire: {
             Normal: 1,
@@ -197,7 +206,8 @@ export const TypeEffectiveness = {
             Ice: 0.5,
             Dragon: 1,
             Dark: 1,
-            Fairy: 0.5
+            Fairy: 0.5,
+            Nuclear: 2
         },
         Water:{
             Normal: 1,
@@ -217,7 +227,8 @@ export const TypeEffectiveness = {
             Ice: 0.5,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Grass:{
             Normal: 1,
@@ -237,7 +248,8 @@ export const TypeEffectiveness = {
             Ice: 2,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Electric:{
             Normal: 1,
@@ -257,7 +269,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Psychic:{
             Normal: 1,
@@ -277,7 +290,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 2,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Ice: {
             Normal: 1,
@@ -297,7 +311,8 @@ export const TypeEffectiveness = {
             Ice: 0.5,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
         },
         Dragon: {
             Normal: 1,
@@ -317,7 +332,8 @@ export const TypeEffectiveness = {
             Ice: 2,
             Dragon: 2,
             Dark: 1,
-            Fairy: 2
+            Fairy: 2,
+            Nuclear: 2
         },
         Dark: {
             Normal: 1,
@@ -337,7 +353,8 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 0.5,
-            Fairy: 2
+            Fairy: 2,
+            Nuclear: 2
         },
         Fairy: {
             Normal: 1,
@@ -357,7 +374,29 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 0,
             Dark: 0.5,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 2
+        },
+        Nuclear: {
+            Normal: 2,
+            Fighting: 2,
+            Flying: 2,
+            Poison: 2,
+            Ground: 2,
+            Rock: 2,
+            Bug: 2,
+            Ghost: 2,
+            Steel: 2,
+            Fire: 2,
+            Water: 2,
+            Grass: 2,
+            Electric: 2,
+            Psychic: 2,
+            Ice: 2,
+            Dragon: 2,
+            Dark: 2,
+            Fairy: 2,
+            Nuclear: 0.5
         },
         null: {
             Normal: 1,
@@ -377,6 +416,7 @@ export const TypeEffectiveness = {
             Ice: 1,
             Dragon: 1,
             Dark: 1,
-            Fairy: 1
+            Fairy: 1,
+            Nuclear: 1
         }
 }